chore(api): move find_machine_health_histories code to handlers
